django_backend/requirements.txt
asgiref==3.7.2
coverage==7.2.7
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==18.13.0
importlib-metadata==6.7.0
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.9
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
zipp==3.15.0
